backend/requirements.txt
fastapi>=0.132.0
uvicorn>=0.41.0
httpx>=0.28.1
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt>=5.0.0
pydantic>=2.12.5
PyJWT>=2.11.0
python-multipart>=0.0.22
requests>=2.32.5
psycopg2-binary>=2.9.9
redis>=5.0.7
sqlalchemy>=2.0.0
email-validator>=2.2.0
aiosmtplib>=3.0.2
google-auth>=2.23.0
cryptography>=42.0.0
